rule dump_fastq: Modified to use an expand() on its output. This allows for dump_fastq to act on all input file and all output files at the same time.
Notes
-----
I believe we will want to split the input in half, because it is being generated twice right now. This is due to the "_1" and "_2" portions incorporated earlier. There is nothing we can do about this for input, but a sorted() input and output will give us the correct matching for creating output with parallel-fastq-dump